% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exerciseGraphs.R
\name{exerciseGraphs}
\alias{exerciseGraphs}
\title{Graphs Based on Member Exercise Data}
\usage{
exerciseGraphs(
  y_axis = data$Workout_Hours.Weekly,
  point_color = data$Workout_Type
)
}
\arguments{
\item{y_axis}{Which variable will be on the x_axis; default = data$Workout_Hours.Weekly}

\item{point_color}{Which variable the points will be colored by; default = data$Workout_Type}
}
\value{
A scatterplot with a regression line
}
\description{
Graphs Based on Member Exercise Data
}
\examples{
exerciseGraphs()
exerciseGraphs(data$Session_Duration.hours,data$Experience_Level)
}
