% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exerciseReg.R
\name{exerciseReg}
\alias{exerciseReg}
\title{Predictive Model based on Member Exercise Data}
\usage{
exerciseReg(x = data$Calories_Burned)
}
\arguments{
\item{x}{Input variable for linear regression to predict BMI; default = data$Calories_Burned}
}
\value{
Plots and summaries of a linear regression model
}
\description{
Predictive Model based on Member Exercise Data
}
\examples{
exerciseReg()
}
