% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendBMI.R
\name{recommendBMI}
\alias{recommendBMI}
\title{Recommend Exercise Routines based on BMI}
\usage{
recommendBMI(x = 1)
}
\arguments{
\item{x}{An index/observation of the "data" data set to select; defaults to 1}
}
\value{
Observations from the "data" data set are printed to the console.
}
\description{
Recommend Exercise Routines based on BMI
}
\examples{
recommendBMI()
recommendBMI(73)
recommendBMI(x = 105)
}
