% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetBMI.R
\name{targetBMI}
\alias{targetBMI}
\title{Recommend Exercise Routine based on a Target BMI}
\usage{
targetBMI(x = 21.75)
}
\arguments{
\item{x}{The BMI you wish to reach; defaults to 21.75 - the middle of the healthy range}
}
\value{
A print out with descriptive text and calculated variables that shows the exercise routine of observations with similar BMI's to the target, x
}
\description{
Recommend Exercise Routine based on a Target BMI
}
\examples{
targetBMI()
targetBMI(28)
targetBMI(x=17)
}
